Refactoring hungyrman package to wherefordinner in notify app.
Created profiles for cloud native only events to remove the RabbitMQ
binder.
