+++ where-for-dinner/where-for-dinner-notify/pom.xml
@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.4</version>
+		<version>2.7.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
@@ -15,7 +15,7 @@
 	<description>Where for Dinner notification application</description>
 	<properties>
 		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.3</spring-cloud.version>
+		<spring-cloud.version>2021.0.4</spring-cloud.version>
 		<snakeyaml.version>1.32</snakeyaml.version>
 	</properties>
 	<dependencies>
@@ -31,14 +31,6 @@
 		    <groupId>org.springframework.boot</groupId>
 		    <artifactId>spring-boot-starter-mail</artifactId>
 		</dependency>	
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream</artifactId>
-		</dependency>	
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
-		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-function-webflux</artifactId>
@@ -59,13 +51,6 @@
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream</artifactId>
-			<scope>test</scope>
-			<classifier>test-binder</classifier>
-			<type>test-jar</type>
-		</dependency>
 	</dependencies>
 	<dependencyManagement>
 		<dependencies>
@@ -78,9 +63,27 @@
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
-
-	<build>
-		<plugins>
+    <repositories>
+        <repository>
+            <id>spring-releases</id>
+            <url>https://repo.spring.io/release/</url>
+            <releases>
+                <enabled>true</enabled>
+            </releases>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
+    </repositories>
+	<pluginRepositories>
+	    <pluginRepository>
+	        <id>spring-release</id>
+	        <name>Spring release</name>
+	        <url>https://repo.spring.io/release</url>
+	    </pluginRepository>
+	</pluginRepositories>
+    <build>
+        <plugins>
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
@@ -93,7 +96,40 @@
 					</excludes>
 				</configuration>
 			</plugin>
-		</plugins>
-	</build>
+        </plugins>
+   </build>
+   <profiles>
+       <profile>
+           <id>streams</id>
+           <activation>
+              <activeByDefault>true</activeByDefault>
+           </activation> 
+           <dependencies>
+			   <dependency>
+		            <groupId>org.springframework.cloud</groupId>
+		            <artifactId>spring-cloud-stream</artifactId>
+		       </dependency>	
+			   <dependency>
+					<groupId>org.springframework.cloud</groupId>
+					<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
+			   </dependency>
+			   <dependency>
+					<groupId>org.springframework.cloud</groupId>
+					<artifactId>spring-cloud-stream</artifactId>
+					<scope>test</scope>
+					<classifier>test-binder</classifier>
+					<type>test-jar</type>
+			   </dependency>
+		   </dependencies>
+       </profile>
+       <profile>
+           <!-- Stubbed for native image build -->
+           <id>native</id>
+           <dependencies>
 
+		   </dependencies>
+           <build>
+           </build>
+       </profile>
+    </profiles>
 </project>
