+++ where-for-dinner/where-for-dinner-notify/src/main/java/com/java/example/tanzu/wherefordinner/config/PublisherConfiguration.java
@@ -1,12 +1,12 @@
-package com.java.example.tanzu.hungryman.config;
+package com.java.example.tanzu.wherefordinner.config;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.mail.javamail.JavaMailSender;
 
-import com.java.example.tanzu.hungryman.publisher.EmailPublisher;
-import com.java.example.tanzu.hungryman.publisher.LoggerPublisher;
+import com.java.example.tanzu.wherefordinner.publisher.EmailPublisher;
+import com.java.example.tanzu.wherefordinner.publisher.LoggerPublisher;
 
 @Configuration
 public class PublisherConfiguration
