+++ where-for-dinner/where-for-dinner-notify/src/main/java/com/java/example/tanzu/wherefordinner/functions/AvailabilitySink.java
@@ -1,4 +1,4 @@
-package com.java.example.tanzu.hungryman.functions;
+package com.java.example.tanzu.wherefordinner.functions;
 
 import java.util.List;
 import java.util.function.Function;
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import com.java.example.tanzu.hungryman.model.Availability;
-import com.java.example.tanzu.hungryman.publisher.Publisher;
+import com.java.example.tanzu.wherefordinner.model.Availability;
+import com.java.example.tanzu.wherefordinner.publisher.Publisher;
 
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
