+++ where-for-dinner/where-for-dinner-notify/src/main/java/com/java/example/tanzu/wherefordinner/model/Availability.java
@@ -1,4 +1,4 @@
-package com.java.example.tanzu.hungryman.model;
+package com.java.example.tanzu.wherefordinner.model;
 
 import java.util.ArrayList;
 import java.util.List;
