+++ where-for-dinner/where-for-dinner-notify/src/main/java/com/java/example/tanzu/wherefordinner/publisher/LoggerPublisher.java
@@ -1,11 +1,11 @@
-package com.java.example.tanzu.hungryman.publisher;
+package com.java.example.tanzu.wherefordinner.publisher;
 
 import java.text.DateFormat;
 import java.time.Instant;
 import java.util.Date;
 import java.util.Locale;
 
-import com.java.example.tanzu.hungryman.model.Availability;
+import com.java.example.tanzu.wherefordinner.model.Availability;
 
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
